+++ src/encoding/evm/group_swaps.rs
