+++ src/encoding/evm/mod.rs
@@ -1,7 +1,8 @@
 pub mod approvals;
 mod constants;
-pub mod encoder_builder;
+pub mod encoder_builders;
+mod group_swaps;
 pub mod strategy_encoder;
 mod swap_encoder;
-pub mod tycho_encoder;
+pub mod tycho_encoders;
 pub mod utils;
