+++ src/encoding/evm/strategy_encoder/mod.rs
@@ -1,3 +1,2 @@
-mod group_swaps;
 pub mod strategy_encoders;
 mod strategy_validators;
