+++ src/encoding/evm/strategy_encoder/strategy_encoders.rs
@@ -12,9 +12,9 @@ use crate::encoding::{
     evm::{
         approvals::permit2::Permit2,
         constants::DEFAULT_ROUTERS_JSON,
-        strategy_encoder::{
-            group_swaps::group_swaps,
-            strategy_validators::{SequentialSwapValidator, SplitSwapValidator, SwapValidator},
+        group_swaps::group_swaps,
+        strategy_encoder::strategy_validators::{
+            SequentialSwapValidator, SplitSwapValidator, SwapValidator,
         },
         swap_encoder::swap_encoder_registry::SwapEncoderRegistry,
         utils::{
@@ -242,8 +242,6 @@ impl SequentialSwapStrategyEncoder {
     }
 }
 
-impl EVMStrategyEncoder for SequentialSwapStrategyEncoder {}
-
 impl StrategyEncoder for SequentialSwapStrategyEncoder {
     fn encode_strategy(&self, solution: Solution) -> Result<(Vec<u8>, Bytes), EncodingError> {
         self.sequential_swap_validator
@@ -303,7 +301,7 @@ impl StrategyEncoder for SequentialSwapStrategyEncoder {
             swaps.push(swap_data);
         }
 
-        let encoded_swaps = self.ple_encode(swaps);
+        let encoded_swaps = ple_encode(swaps);
         let method_calldata = if let Some(permit2) = self.permit2.clone() {
             let (permit, signature) = permit2.get_permit(
                 &self.router_address,
@@ -594,77 +592,6 @@ impl StrategyEncoder for SplitSwapStrategyEncoder {
     }
 }
 
-/// This strategy encoder is used for solutions that are sent directly to the executor, bypassing
-/// the router. Only one solution with one swap is supported.
-///
-/// # Fields
-/// * `swap_encoder_registry`: SwapEncoderRegistry, containing all possible swap encoders
-#[derive(Clone)]
-pub struct ExecutorStrategyEncoder {
-    swap_encoder_registry: SwapEncoderRegistry,
-}
-
-impl ExecutorStrategyEncoder {
-    pub fn new(swap_encoder_registry: SwapEncoderRegistry) -> Self {
-        Self { swap_encoder_registry }
-    }
-}
-
-impl StrategyEncoder for ExecutorStrategyEncoder {
-    fn encode_strategy(&self, solution: Solution) -> Result<(Vec<u8>, Bytes), EncodingError> {
-        let grouped_swaps = group_swaps(solution.clone().swaps);
-        let number_of_groups = grouped_swaps.len();
-        if number_of_groups > 1 {
-            return Err(EncodingError::InvalidInput(format!(
-                "Executor strategy only supports one swap for non-groupable protocols. Found {}",
-                number_of_groups
-            )))
-        }
-
-        let grouped_swap = grouped_swaps
-            .first()
-            .ok_or_else(|| EncodingError::FatalError("Swap grouping failed".to_string()))?;
-
-        let receiver = solution.receiver;
-
-        let swap_encoder = self
-            .get_swap_encoder(&grouped_swap.protocol_system)
-            .ok_or_else(|| {
-                EncodingError::InvalidInput(format!(
-                    "Swap encoder not found for protocol: {}",
-                    grouped_swap.protocol_system
-                ))
-            })?;
-
-        let mut grouped_protocol_data: Vec<u8> = vec![];
-        for swap in grouped_swap.swaps.iter() {
-            let encoding_context = EncodingContext {
-                receiver: receiver.clone(),
-                exact_out: solution.exact_out,
-                router_address: None,
-                group_token_in: grouped_swap.input_token.clone(),
-                group_token_out: grouped_swap.output_token.clone(),
-            };
-            let protocol_data = swap_encoder.encode_swap(swap.clone(), encoding_context.clone())?;
-            grouped_protocol_data.extend(protocol_data);
-        }
-
-        let executor_address = Bytes::from_str(swap_encoder.executor_address())
-            .map_err(|_| EncodingError::FatalError("Invalid executor address".to_string()))?;
-
-        Ok((grouped_protocol_data, executor_address))
-    }
-
-    fn get_swap_encoder(&self, protocol_system: &str) -> Option<&Box<dyn SwapEncoder>> {
-        self.swap_encoder_registry
-            .get_encoder(protocol_system)
-    }
-
-    fn clone_box(&self) -> Box<dyn StrategyEncoder> {
-        Box::new(self.clone())
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::{collections::HashMap, str::FromStr};
@@ -697,198 +624,6 @@ mod tests {
         let eth_chain = eth_chain();
         SwapEncoderRegistry::new(Some("config/test_executor_addresses.json".to_string()), eth_chain)
             .unwrap()
-    }
-
-    #[test]
-    fn test_executor_strategy_encode() {
-        let swap_encoder_registry = get_swap_encoder_registry();
-        let encoder = ExecutorStrategyEncoder::new(swap_encoder_registry);
-
-        let token_in = weth();
-        let token_out = Bytes::from("0x6b175474e89094c44da98b954eedeac495271d0f");
-
-        let swap = Swap {
-            component: ProtocolComponent {
-                id: "0xA478c2975Ab1Ea89e8196811F51A7B7Ade33eB11".to_string(),
-                protocol_system: "uniswap_v2".to_string(),
-                ..Default::default()
-            },
-            token_in: token_in.clone(),
-            token_out: token_out.clone(),
-            split: 0f64,
-        };
-
-        let solution = Solution {
-            exact_out: false,
-            given_token: token_in,
-            given_amount: BigUint::from(1000000000000000000u64),
-            expected_amount: Some(BigUint::from(1000000000000000000u64)),
-            checked_token: token_out,
-            checked_amount: None,
-            sender: Bytes::from_str("0x0000000000000000000000000000000000000000").unwrap(),
-            // The receiver was generated with `makeAddr("bob") using forge`
-            receiver: Bytes::from_str("0x1d96f2f6bef1202e4ce1ff6dad0c2cb002861d3e").unwrap(),
-            swaps: vec![swap],
-            slippage: None,
-            native_action: None,
-        };
-
-        let (protocol_data, executor_address) = encoder
-            .encode_strategy(solution)
-            .unwrap();
-        let hex_protocol_data = encode(&protocol_data);
-        assert_eq!(
-            executor_address,
-            Bytes::from_str("0x5615deb798bb3e4dfa0139dfa1b3d433cc23b72f").unwrap()
-        );
-        assert_eq!(
-            hex_protocol_data,
-            String::from(concat!(
-                // in token
-                "c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2",
-                // component id
-                "a478c2975ab1ea89e8196811f51a7b7ade33eb11",
-                // receiver
-                "1d96f2f6bef1202e4ce1ff6dad0c2cb002861d3e",
-                // zero for one
-                "00",
-            ))
-        );
-    }
-
-    #[test]
-    fn test_executor_strategy_encode_too_many_swaps() {
-        let swap_encoder_registry = get_swap_encoder_registry();
-        let encoder = ExecutorStrategyEncoder::new(swap_encoder_registry);
-
-        let token_in = weth();
-        let token_out = Bytes::from("0x6b175474e89094c44da98b954eedeac495271d0f");
-
-        let swap = Swap {
-            component: ProtocolComponent {
-                protocol_system: "uniswap_v2".to_string(),
-                ..Default::default()
-            },
-            token_in: token_in.clone(),
-            token_out: token_out.clone(),
-            split: 0f64,
-        };
-
-        let solution = Solution {
-            exact_out: false,
-            given_token: token_in,
-            given_amount: BigUint::from(1000000000000000000u64),
-            expected_amount: Some(BigUint::from(1000000000000000000u64)),
-            checked_token: token_out,
-            checked_amount: None,
-            sender: Bytes::from_str("0x0000000000000000000000000000000000000000").unwrap(),
-            receiver: Bytes::from_str("0x1d96f2f6bef1202e4ce1ff6dad0c2cb002861d3e").unwrap(),
-            swaps: vec![swap.clone(), swap],
-            slippage: None,
-            native_action: None,
-        };
-
-        let result = encoder.encode_strategy(solution);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_executor_strategy_encode_grouped_swaps() {
-        let swap_encoder_registry = get_swap_encoder_registry();
-        let encoder = ExecutorStrategyEncoder::new(swap_encoder_registry);
-
-        let eth = eth();
-        let usdc = Bytes::from_str("0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48").unwrap();
-        let pepe = Bytes::from_str("0x6982508145454Ce325dDbE47a25d4ec3d2311933").unwrap();
-
-        // Fee and tick spacing information for this test is obtained by querying the
-        // USV4 Position Manager contract: 0xbd216513d74c8cf14cf4747e6aaa6420ff64ee9e
-        // Using the poolKeys function with the first 25 bytes of the pool id
-        let pool_fee_usdc_eth = Bytes::from(BigInt::from(3000).to_signed_bytes_be());
-        let tick_spacing_usdc_eth = Bytes::from(BigInt::from(60).to_signed_bytes_be());
-        let mut static_attributes_usdc_eth: HashMap<String, Bytes> = HashMap::new();
-        static_attributes_usdc_eth.insert("key_lp_fee".into(), pool_fee_usdc_eth);
-        static_attributes_usdc_eth.insert("tick_spacing".into(), tick_spacing_usdc_eth);
-
-        let pool_fee_eth_pepe = Bytes::from(BigInt::from(25000).to_signed_bytes_be());
-        let tick_spacing_eth_pepe = Bytes::from(BigInt::from(500).to_signed_bytes_be());
-        let mut static_attributes_eth_pepe: HashMap<String, Bytes> = HashMap::new();
-        static_attributes_eth_pepe.insert("key_lp_fee".into(), pool_fee_eth_pepe);
-        static_attributes_eth_pepe.insert("tick_spacing".into(), tick_spacing_eth_pepe);
-
-        let swap_usdc_eth = Swap {
-            component: ProtocolComponent {
-                id: "0xdce6394339af00981949f5f3baf27e3610c76326a700af57e4b3e3ae4977f78d"
-                    .to_string(),
-                protocol_system: "uniswap_v4".to_string(),
-                static_attributes: static_attributes_usdc_eth,
-                ..Default::default()
-            },
-            token_in: usdc.clone(),
-            token_out: eth.clone(),
-            split: 0f64,
-        };
-
-        let swap_eth_pepe = Swap {
-            component: ProtocolComponent {
-                id: "0xecd73ecbf77219f21f129c8836d5d686bbc27d264742ddad620500e3e548e2c9"
-                    .to_string(),
-                protocol_system: "uniswap_v4".to_string(),
-                static_attributes: static_attributes_eth_pepe,
-                ..Default::default()
-            },
-            token_in: eth.clone(),
-            token_out: pepe.clone(),
-            split: 0f64,
-        };
-
-        let solution = Solution {
-            exact_out: false,
-            given_token: usdc,
-            given_amount: BigUint::from_str("1000_000000").unwrap(),
-            checked_token: pepe,
-            expected_amount: Some(BigUint::from_str("105_152_000000000000000000").unwrap()),
-            checked_amount: None,
-            slippage: None,
-            sender: Bytes::from_str("0xcd09f75E2BF2A4d11F3AB23f1389FcC1621c0cc2").unwrap(),
-            receiver: Bytes::from_str("0xcd09f75E2BF2A4d11F3AB23f1389FcC1621c0cc2").unwrap(),
-            swaps: vec![swap_usdc_eth, swap_eth_pepe],
-            ..Default::default()
-        };
-
-        let (protocol_data, executor_address) = encoder
-            .encode_strategy(solution)
-            .unwrap();
-        let hex_protocol_data = encode(&protocol_data);
-        assert_eq!(
-            executor_address,
-            Bytes::from_str("0xf62849f9a0b5bf2913b396098f7c7019b51a820a").unwrap()
-        );
-        assert_eq!(
-            hex_protocol_data,
-            String::from(concat!(
-                // group in token
-                "a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48",
-                // group out token
-                "6982508145454ce325ddbe47a25d4ec3d2311933",
-                // zero for one
-                "00",
-                // executor address
-                "f62849f9a0b5bf2913b396098f7c7019b51a820a",
-                // first pool intermediary token (ETH)
-                "0000000000000000000000000000000000000000",
-                // fee
-                "000bb8",
-                // tick spacing
-                "00003c",
-                // second pool intermediary token (PEPE)
-                "6982508145454ce325ddbe47a25d4ec3d2311933",
-                // fee
-                "0061a8",
-                // tick spacing
-                "0001f4"
-            ))
-        );
     }
 
     #[rstest]
